🟢 index.tsx for new files 🛠️ index.tsx -> added login form with validation and submission logic 🟢 login.module.css for new files 🟢 index.tsx for new files 🛠️ index.tsx -> added signup form with validation and submission logic 🟢 signup.module.css for new files 🔴 login.css
