+++ qwik/src/components/auth/signup/index.tsx
@@ -1,0 +1,88 @@
+import { component$, useSignal, $ } from "@builder.io/qwik";
+import styles from "./signup.module.css";
+
+export default component$(() => {
+  const username = useSignal('');
+  const email = useSignal('');
+  const password = useSignal('');
+  const successMessage = useSignal<string | null>(null);
+  const errorMessage = useSignal<string | null>(null);
+
+  const handleSubmit = $(async (event: Event) => {
+    event.preventDefault();
+    errorMessage.value = null;
+    successMessage.value = null;
+
+    try {
+      const response = await fetch("https://api.kspcommunity.com/api/signup", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          username: username.value,
+          email: email.value,
+          password: password.value,
+        }),
+      });
+
+      if (!response.ok) {
+        throw new Error("Failed to sign up. Please try again.");
+      }
+
+      successMessage.value = "Signup successful!";
+      username.value = '';
+      email.value = '';
+      password.value = '';
+    } catch (error) {
+      errorMessage.value = (error as Error).message;
+    }
+  });
+
+  return (
+    <>
+      <div class={[styles.container, styles["container-center"]].join(' ')}>
+        <p class={styles.heading}>
+          Sign <span class={styles.highlight}>Up</span>
+        </p>
+        {errorMessage.value && <pre class={styles.error}>{errorMessage.value}</pre>}
+        {successMessage.value && <pre class={styles.success}>{successMessage.value}</pre>}
+        <form onSubmit$={handleSubmit} class={styles["signup-form"]}>
+          <div>
+            <label for="username">Username:</label>
+            <input
+              id="username"
+              type="text"
+              value={username.value}
+              onInput$={(e) => (username.value = (e.target as HTMLInputElement).value)}
+              required
+            />
+          </div>
+          <div>
+            <label for="email">Email:</label>
+            <input
+              id="email"
+              type="email"
+              value={email.value}
+              onInput$={(e) => (email.value = (e.target as HTMLInputElement).value)}
+              required
+            />
+          </div>
+          <div>
+            <label for="password">Password:</label>
+            <input
+              id="password"
+              type="password"
+              value={password.value}
+              onInput$={(e) => (password.value = (e.target as HTMLInputElement).value)}
+              required
+            />
+          </div>
+          <div>
+            <button type="submit">Sign Up</button>
+          </div>
+        </form>
+      </div>
+    </>
+  );
+});
